+++ packages/box/src/components/Box.tsx
@@ -4,11 +4,14 @@ import getBoxWidth from '../methods/getBoxWidth'
 import BoxPadding from '../interfaces/BoxPadding'
 import getBoxPaddingX from '../methods/getBoxPaddingX'
 import getBoxPaddingY from '../methods/getBoxPaddingY'
+import BoxAlign from '../interfaces/BoxAlign'
+import getBoxAlign from '../methods/getBoxAlign'
 
 interface Props extends HTMLProps<HTMLDivElement> {
   width?: BoxWidth | 'no'
   px?: BoxPadding | 'no'
   py?: BoxPadding | 'no'
+  align?: BoxAlign | 'no'
   children?: ReactNode | ReactNode[]
 }
 
@@ -16,6 +19,7 @@ export default function Box({
   width = 'no',
   px = 'no',
   py = 'no',
+  align = 'no',
   children,
   className,
   ...props
@@ -23,12 +27,13 @@ export default function Box({
   let widthClasses = getBoxWidth(width)
   let paddingXClasses = getBoxPaddingX(px)
   let paddingYClasses = getBoxPaddingY(py)
+  let alignClasses = getBoxAlign(align)
+
+  const classNames = [widthClasses, paddingXClasses, paddingYClasses, alignClasses, className].filter(Boolean).join(' ').trim()
 
   return (
     <div
-      className={`${widthClasses} ${paddingXClasses} ${paddingYClasses} ${
-        className || ''
-      }`}
+      className={classNames}
       {...props}
     >
       {children}