+++ packages/box/src/interfaces/BoxAlign.ts
@@ -1,0 +1,3 @@
+type BoxAlign = 'left' | 'center' | 'right'
+
+export default BoxAlign