+++ packages/box/src/interfaces/BoxWidth.ts
@@ -1,3 +1,3 @@
-type BoxWidth = 'small' | 'medium' | 'large' | 'extra' | 'full' | 'content'
+type BoxWidth = 'small' | 'medium' | 'large' | 'extra' | 'full' | 'content-small' | 'content' | 'content-large'
 
 export default BoxWidth