+++ packages/box/src/methods/getBoxWidth.ts
@@ -5,22 +5,28 @@ export default function getBoxWidth(width: BoxWidth | 'no') {
 
   switch (width) {
     case 'small':
-      className = 'max-w-sm w-full'
+      className = 'max-w-[500px] w-full'
       break
     case 'medium':
-      className = 'max-w-md w-full'
+      className = 'max-w-[800px] w-full'
       break
     case 'large':
-      className = 'max-w-lg w-full'
+      className = 'max-w-[1100px] w-full'
       break
     case 'extra':
-      className = 'max-w-xl w-full'
+      className = 'max-w-[1300px] w-full'
       break
     case 'full':
       className = 'w-full'
       break
+    case 'content-small':
+      className = 'max-w-[600px] w-full'
+      break
     case 'content':
-      className = 'max-w-[750px] w-full'
+      className = 'max-w-[650px] w-full'
+      break
+    case 'content-large':
+      className = 'max-w-[700px] w-full'
       break
     case 'no':
       className = ''