+++ packages/box/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2024-05-25
+
+### Breaking
+
+- Changed the width's values of all sizes to be more useful and visually pleasing.
+
+### Added
+
+- Added the new `align` prop to align the box either to the left, center, or right.
+
 ## [1.0.0] - 2024-05-10
 
 ### Added